apply changes on vs-bar, vs-header, vs-footer
