--- packages/vlossom/src/components/vs-bar/index.ts
@@ -1,10 +1,0 @@
-import type VsBar from './VsBar.vue';
-
-declare module 'vue' {
-    interface GlobalComponents {
-        VsBar: typeof VsBar;
-    }
-}
-
-export * from './types';
-export type { VsBar };
