--- packages/vlossom/src/components/vs-footer/index.ts
@@ -1,10 +1,0 @@
-import type VsFooter from './VsFooter.vue';
-
-declare module 'vue' {
-    interface GlobalComponents {
-        VsFooter: typeof VsFooter;
-    }
-}
-
-export * from './types';
-export type { VsFooter };
