+++ packages/vlossom/src/components/vs-footer/types.ts
@@ -1,3 +1,12 @@
+import type VsFooter from './VsFooter.vue';
 import type { VsBarStyleSet } from '@/components/vs-bar/types';
+
+declare module 'vue' {
+    interface GlobalComponents {
+        VsFooter: typeof VsFooter;
+    }
+}
+
+export type { VsFooter };
 
 export interface VsFooterStyleSet extends VsBarStyleSet {}
