--- packages/vlossom/src/components/vs-header/index.ts
@@ -1,10 +1,0 @@
-import type VsHeader from './VsHeader.vue';
-
-declare module 'vue' {
-    interface GlobalComponents {
-        VsHeader: typeof VsHeader;
-    }
-}
-
-export * from './types';
-export type { VsHeader };
